Started Work On New Quest System (Tasks)
Task_Manager stores active and completed tasks. It also runs logic for those tasks.
Control_Task allows other scripts to activate tasks, complete tasks, and modify requirements of tasks.
GameTask is the base class for tasks (quests).
TaskRequirement is the base of logic for tasks, stored as an integer, when less than one, is is true. When all requirements on a task are true, the task is completed.